requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Google Cloud and AI services
google-cloud-firestore==2.13.1
google-generativeai==0.3.2

# Trend analysis
pytrends==4.9.2
praw==7.7.1

# Image services
httpx==0.25.2
requests==2.31.0

# Data validation and configuration
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging
structlog==23.2.0